receive ACK/NAK on single byte reads
Notes from demo apps, building blocks in smaller pieces. This is both
a) way slower, and
b) not as reliable.  The stop bit generation isn't reliably enough.

Further, only being able to send a single byte to get the ack is pretty
untenable.  Perhaps you can use the ack detection on the writing of the
address, then use the normal multi command patterns to get some decent
performance, and hope it works ok?